Navigation: Updated styling and some reorganization (tbd.)
