+++ .storybook/preview-head.html
@@ -1,3 +1,8 @@
 <style>
     @import url('https://fonts.googleapis.com/css?family=Open+Sans:400,600|Titillium+Web:400,600');
+
+    body {
+        font-family: var(--font-family-primary);
+        font-size: var(--font-size-normal);
+    }
 </style>
