+++ src/components/Navigation/NavBar.scss
@@ -1,7 +1,7 @@
 @import '../../themes/oneui/oneui.scss';
 
 .NavBar {
-    background-color: var(--color-neutral);
+    background-color: var(--color-neutral-10);
     color: var(--color-brand);
     width: 100%;
     height: var(--navigation-height);
@@ -17,9 +17,5 @@
         justify-content: flex-start;
         align-items: flex-start;
         align-content: center;
-    }
-
-    a {
-        text-decoration: none;
     }
 }
