+++ src/components/Navigation/NavItem/NavItem.scss
@@ -1,0 +1,36 @@
+@import '../../../themes/oneui/oneui.scss';
+
+.NavItem {
+    color: var(--link-color-normal);
+    display: flex;
+    align-items: center;
+    height: 100%;
+    font: {
+        size: var(--font-size-large);
+        weight: var(--font-weight-bold);
+    }
+    box-sizing: border-box;
+    padding: 0 var(--spacing-5x);
+    transition: all var(--transition-duration);
+    text-decoration: none;
+
+    &:hover {
+        color: var(--link-color-hover);
+        box-shadow: inset 0 -3px 0 var(--color-neutral-25);
+    }
+
+    &__active,
+    &--active {
+        color: var(--color-foreground);
+        box-shadow: inset 0 -3px 0 var(--color-primary);
+
+        &:hover {
+            color: var(--color-foreground);
+            box-shadow: inset 0 -3px 0 var(--color-primary-40);
+        }
+    }
+
+    &--pullRight {
+        margin-left: auto;
+    }
+}
