+++ src/components/Navigation/NavItem/index.js
@@ -1,0 +1,1 @@
+export { default } from './NavItem';
