+++ src/components/Navigation/index.js
@@ -1,3 +1,1 @@
-export { default as NavBar } from './NavBar';
-export { default as NavItem } from './NavItem';
-export { default as NavLinkItem } from './NavLinkItem';
+export { default } from './NavBar';
