+++ src/index.js
@@ -10,7 +10,9 @@ export { default as Heading } from './components/Heading';
 export { default as Input } from './components/Input';
 export { default as Link } from './components/Link';
 export { default as LoadingSpinner } from './components/LoadingSpinner';
-export * from './components/Navigation';
+export { default as NavItem } from './components/Navigation/NavItem';
+export { default as NavLinkItem } from './components/Navigation/NavLinkItem';
+export { default as NavBar } from './components/Navigation';
 export { default as ProgressBar } from './components/ProgressBar';
 export { default as RadioButton } from './components/RadioButton';
 export { default as Tab } from './components/Tabs/Tab';
