+++ src/themes/oneui/base-variables/_colors.scss
@@ -42,7 +42,7 @@ $color-bad:         #F42534 !default;
         warning: $color-warning,
         bad: $color-bad
     ) {
-        @each $factor in (20, 50) {
+        @each $factor in (20, 50, 80) {
             --color-#{ $context }-#{ 50 - ($factor / 2) }: #{ mix($color-background, $value, $factor) };
             --color-#{ $context }-#{ 50 + ($factor / 2) }: #{ mix($color-foreground, $value, $factor) };
         }
